feat(#227): 서버 scale-out을 고려하여 알림을 전송하는 NotificationManager 추상화
