+++ server/src/main/java/com/hexacore/tayo/reservation/ReservationController.java
@@ -2,7 +2,7 @@ package com.hexacore.tayo.reservation;
 
 import com.hexacore.tayo.common.errors.GeneralException;
 import com.hexacore.tayo.common.response.Response;
-import com.hexacore.tayo.notification.NotificationManager;
+import com.hexacore.tayo.notification.manager.NotificationManager;
 import com.hexacore.tayo.notification.model.NotificationType;
 import com.hexacore.tayo.reservation.dto.CreateReservationRequestDto;
 import com.hexacore.tayo.reservation.dto.CreateReservationResponseDto;
@@ -36,9 +36,9 @@ public class ReservationController {
 
     @PostMapping
     public ResponseEntity<Response> createReservation(HttpServletRequest request,
-            @Valid @RequestParam String orderName,
-            @Valid @RequestParam String userName,
-            @Valid @RequestBody CreateReservationRequestDto createReservationRequestDto) {
+                                                      @Valid @RequestParam String orderName,
+                                                      @Valid @RequestParam String userName,
+                                                      @Valid @RequestBody CreateReservationRequestDto createReservationRequestDto) {
         Long guestUserId = (Long) request.getAttribute("userId");
 
         // 예약 진행: DB 업데이트
@@ -81,8 +81,8 @@ public class ReservationController {
 
     @PatchMapping("/{reservationId}")
     public ResponseEntity<Response> updateReservationStatus(HttpServletRequest request,
-            @PathVariable Long reservationId,
-            @RequestBody UpdateReservationStatusRequestDto statusDto) {
+                                                            @PathVariable Long reservationId,
+                                                            @RequestBody UpdateReservationStatusRequestDto statusDto) {
         Long userId = (Long) request.getAttribute("userId");
         String status = statusDto.getStatus();
